Update the Workout Details and delete functionality using useAuth context

+++ front-end/src/components/WorkoutDetails.js
@@ -2,13 +2,22 @@ import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 
 //date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      console.log("You must be logged in");
+      return;
+    }
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
